refactor(client-presence): split files for encapsulated support 1. Separate types from implementation files
   origin -> new types location
   -------------------------------
   latestValueManager.ts -> latestTypes.ts
   latestMapValueManager.ts -> latestMapTypes.ts

2. Separate implementation `createValidatedGetter` helper from types file

   origin -> new location
   -----------------------
   latestValueTypes.ts -> validatedGetter.ts

3. Cleanup `StateFactory` doc generation

   - Import the types required by api-extractor to generate docs without internal file listings.
   - Revise `link` references to not directly specify properties. Break up into `StateFactory` reference and reference to the respective factory type.
